refactored profile frontend
